PROJECT 1 - Books/requirements.txt
asgiref==3.8.1
Django==5.1.11
djangorestframework==3.15.2
sqlparse==0.5.3
